osdp: type enum-valued struct fields as their enum
Several public struct fields whose domain is an enum were declared
uint8_t: the biometric status/type/format fields, the PD capability
function_code, and the LED on/off colors. Give each its enum type,
matching how the other control-code and format fields are declared.

The Python marshaling table encodes field width independently, so
retype the affected descriptors from U8_FIELD to INT_FIELD_ and add
ASSERT_INT_SIZED guards. Cast the now-enum function_code in the
signed capability comparisons to silence -Wsign-compare.

led_params.control_code stays uint8_t: the temporary and permanent
blocks share the struct but their control-code enums disagree.

+++ src/osdp_cp.c
@@ -469,7 +469,7 @@ static int cp_decode_response(struct osdp_pd *pd, uint8_t *buf, int len)
 
 		/* Get peer RX buffer size */
 		t = OSDP_PD_CAP_RECEIVE_BUFFERSIZE;
-		if (pd->cap[t].function_code == t) {
+		if ((int)pd->cap[t].function_code == t) {
 			pd->peer_rx_size = pd->cap[t].compliance_level;
 			pd->peer_rx_size |= (uint32_t)pd->cap[t].num_items << 8;
 		}
@@ -484,7 +484,7 @@ static int cp_decode_response(struct osdp_pd *pd, uint8_t *buf, int len)
 
 		/* Check checksum/CRC support capability */
 		t = OSDP_PD_CAP_CHECK_CHARACTER_SUPPORT;
-		if (pd->cap[t].function_code == t) {
+		if ((int)pd->cap[t].function_code == t) {
 			if (pd->cap[t].compliance_level & 0x01) {
 				SET_FLAG(pd, PD_FLAG_CP_USE_CRC);
 			} else {
